+++ Sources/ZcashLightClientKit/Block/Processor/CompactBlockProcessor.swift
@@ -30,6 +30,7 @@ public enum CompactBlockProcessorError: Error {
     case wrongConsensusBranchId(expectedLocally: ConsensusBranchID, found: ConsensusBranchID)
     case networkMismatch(expected: NetworkType, found: NetworkType)
     case saplingActivationMismatch(expected: BlockHeight, found: BlockHeight)
+    case unknown
 }
 
 /**
@@ -311,11 +312,30 @@ public class CompactBlockProcessor {
         case synced
     }
     
-    public private(set) var state: State = .stopped {
-        didSet {
-            transitionState(from: oldValue, to: self.state)
+    // TODO: this isn't an Actor even though it looks like a good candidate, the reason:
+    // `state` lives in both sync and async environments. An Actor is demanding async context only
+    // so we can't take the advantage unless we encapsulate all `state` reads/writes to async context.
+    // Therefore solution with class + lock works for us butr eventually will be replaced.
+    // The future of CompactBlockProcessor is an actor (we won't need to encapsulate the state separately), issue 523,
+    // https://github.com/zcash/ZcashLightClientKit/issues/523
+    public class ThreadSafeState {
+        private var state: State = .stopped
+        let lock = NSLock()
+        
+        func setState(_ newState: State) {
+            lock.lock()
+            defer { lock.unlock() }
+            state = newState
+        }
+        
+        public func getState() -> State {
+            lock.lock()
+            defer { lock.unlock() }
+            return state
         }
     }
+    
+    public internal(set) var state = ThreadSafeState()
 
     var config: Configuration {
         willSet {
@@ -328,7 +348,7 @@ public class CompactBlockProcessor {
     }
 
     var shouldStart: Bool {
-        switch self.state {
+        switch self.state.getState() {
         case .stopped, .synced, .error:
             return !maxAttemptsReached
         default:
@@ -336,19 +356,19 @@ public class CompactBlockProcessor {
         }
     }
 
-    private var service: LightWalletService
+    var service: LightWalletService
     private(set) var downloader: CompactBlockDownloading
-    private var storage: CompactBlockStorage
-    private var transactionRepository: TransactionRepository
-    private var accountRepository: AccountRepository
-    private var rustBackend: ZcashRustBackendWelding.Type
+    var storage: CompactBlockStorage
+    var transactionRepository: TransactionRepository
+    var accountRepository: AccountRepository
+    var rustBackend: ZcashRustBackendWelding.Type
     private var retryAttempts: Int = 0
     private var backoffTimer: Timer?
     private var lowerBoundHeight: BlockHeight?
     private var latestBlockHeight: BlockHeight
     private var lastChainValidationFailure: BlockHeight?
     private var consecutiveChainValidationErrors: Int = 0
-    private var processingError: Error?
+    var processingError: Error?
     private var foundBlocks = false
     private var maxAttempts: Int {
         config.retries
@@ -358,13 +378,7 @@ public class CompactBlockProcessor {
         BlockHeight(self.config.downloadBatchSize)
     }
 
-    private var operationQueue: OperationQueue = {
-        let queue = OperationQueue()
-        queue.name = "CompactBlockProcessorQueue"
-        queue.maxConcurrentOperationCount = 1
-        return queue
-    }()
-
+    private var cancelableTask: Task<Void, Error>?
 
     /// Initializes a CompactBlockProcessor instance
     /// - Parameters:
@@ -431,7 +445,13 @@ public class CompactBlockProcessor {
     }
     
     deinit {
-        self.operationQueue.cancelAllOperations()
+        cancelableTask?.cancel()
+    }
+
+    func setState(_ newState: State) {
+        let oldValue = state.getState()
+        state.setState(newState)
+        transitionState(from: oldValue, to: newState)
     }
 
     static func validateServerInfo(
@@ -487,19 +507,14 @@ public class CompactBlockProcessor {
             self.backoffTimer?.invalidate()
             self.backoffTimer = nil
         }
-        guard !operationQueue.isSuspended else {
-            LoggerProxy.debug("restarting suspended queue")
-            operationQueue.isSuspended = false
-            return
-        }
 
         guard shouldStart else {
-            switch self.state {
+            switch self.state.getState() {
             case .error(let e):
                 // max attempts have been reached
                 LoggerProxy.info("max retry attempts reached with error: \(e)")
                 notifyError(CompactBlockProcessorError.maxAttemptsReached(attempts: self.maxAttempts))
-                self.state = .stopped
+                setState(.stopped)
             case .stopped:
                 // max attempts have been reached
                 LoggerProxy.info("max retry attempts reached")
@@ -514,7 +529,7 @@ public class CompactBlockProcessor {
             return
         }
 
-        self.nextBatchTask()
+        self.nextBatch()
     }
 
     /**
@@ -523,18 +538,14 @@ public class CompactBlockProcessor {
     Note: retry count is reset
     - Parameter cancelTasks: cancel the pending tasks. Defaults to true
     */
-    public func stop(cancelTasks: Bool = true) {
+    public func stop() {
         self.backoffTimer?.invalidate()
         self.backoffTimer = nil
 
-        if cancelTasks {
-            operationQueue.cancelAllOperations()
-        } else {
-            self.operationQueue.isSuspended = true
-        }
+        cancelableTask?.cancel()
 
         self.retryAttempts = 0
-        self.state = .stopped
+        setState(.stopped)
     }
 
     /**
@@ -577,7 +588,7 @@ public class CompactBlockProcessor {
     - Throws CompactBlockProcessorError.invalidConfiguration if block height is invalid or if processor is already started
     */
     func setStartHeight(_ startHeight: BlockHeight) throws {
-        guard self.state == .stopped, startHeight >= config.network.constants.saplingActivationHeight else {
+        guard self.state.getState() == .stopped, startHeight >= config.network.constants.saplingActivationHeight else {
             throw CompactBlockProcessorError.invalidConfiguration
         }
 
@@ -612,243 +623,34 @@ public class CompactBlockProcessor {
         })
     }
 
-    /**
-    processes new blocks on the given range based on the configuration set for this instance
-    the way operations are queued is implemented based on the following good practice https://forums.developer.apple.com/thread/25761
-     
-    */
-    // swiftlint:disable cyclomatic_complexity
+    
+    /// Processes new blocks on the given range based on the configuration set for this instance
     func processNewBlocks(range: CompactBlockRange) {
         self.foundBlocks = true
         self.backoffTimer?.invalidate()
         self.backoffTimer = nil
         
-        let cfg = self.config
-        let downloadBlockOperation = CompactBlockStreamDownloadOperation(
-            service: self.service,
-            storage: self.storage,
-            blockBufferSize: self.config.downloadBufferSize,
-            startHeight: range.lowerBound,
-            targetHeight: range.upperBound,
-            progressDelegate: self
-        )
-        
-        downloadBlockOperation.startedHandler = { [weak self] in
-            DispatchQueue.main.async {
-                self?.state = .downloading
-            }
-        }
-        
-        downloadBlockOperation.errorHandler = { [weak self] error in
-            DispatchQueue.main.async {
-                guard let self = self else { return }
-                self.processingError = error
-                self.fail(error)
-            }
-        }
-        
-        let validateChainOperation = CompactBlockValidationOperation(
-            rustWelding: self.rustBackend,
-            cacheDb: cfg.cacheDb,
-            dataDb: cfg.dataDb,
-            networkType: self.config.network.networkType
-        )
-        
-        let downloadValidateAdapterOperation = BlockOperation { [weak validateChainOperation, weak downloadBlockOperation] in
-            validateChainOperation?.error = downloadBlockOperation?.error
-        }
-        
-        validateChainOperation.completionHandler = { [weak self] _, cancelled in
-            guard !cancelled else {
-                DispatchQueue.main.async {
-                    self?.state = .stopped
-                    LoggerProxy.debug("Warning: validateChainOperation operation cancelled")
-                }
-                return
-            }
-            
-            LoggerProxy.debug("validateChainFinished")
-        }
-        
-        validateChainOperation.errorHandler = { [weak self] error in
-            DispatchQueue.main.async { [weak self] in
-                guard let self = self else { return }
-
-                guard let validationError = error as? CompactBlockValidationError else {
-                    LoggerProxy.error("Warning: validateChain operation returning generic error: \(error)")
-                    return
-                }
-                
-                switch validationError {
-                case .validationFailed(let height):
-                    LoggerProxy.debug("chain validation at height: \(height)")
-                    self.validationFailed(at: height)
-                case .failedWithError(let e):
-                    guard let validationFailure = e else {
-                        LoggerProxy.error("validation failed without a specific error")
-                        self.fail(CompactBlockProcessorError.generalError(message: "validation failed without a specific error"))
-                        return
-                    }
+        cancelableTask = Task(priority: .userInitiated) {
+            do {
+                try await compactBlockStreamDownload(
+                    blockBufferSize: config.downloadBufferSize,
+                    startHeight: range.lowerBound,
+                    targetHeight: range.upperBound
+                )
+                try await compactBlockValidation()
+                try await compactBlockBatchScanning(range: range)
+                try await compactBlockEnhancement(range: range)
+                try await fetchUnspentTxOutputs(range: range)
+            } catch {
+                if error is CancellationError {
                     
-                    self.fail(validationFailure)
+                }
+                
+                if !(Task.isCancelled) {
+                    fail(error)
                 }
             }
         }
-        
-        validateChainOperation.startedHandler = { [weak self] in
-            DispatchQueue.main.async { [weak self] in
-                self?.state = .validating
-            }
-        }
-        
-        let scanBlocksOperation = CompactBlockBatchScanningOperation(
-            rustWelding: rustBackend,
-            cacheDb: config.cacheDb,
-            dataDb: config.dataDb,
-            transactionRepository: transactionRepository,
-            range: range,
-            batchSize: UInt32(self.config.scanningBatchSize),
-            networkType: self.config.network.networkType,
-            progressDelegate: self
-        )
-        
-        let validateScanningAdapterOperation = BlockOperation { [weak scanBlocksOperation, weak validateChainOperation] in
-            scanBlocksOperation?.error = validateChainOperation?.error
-        }
-
-        scanBlocksOperation.startedHandler = { [weak self] in
-            DispatchQueue.main.async { [weak self] in
-                self?.state = .scanning
-            }
-        }
-        
-        scanBlocksOperation.completionHandler = { [weak self] _, cancelled in
-            guard !cancelled else {
-                DispatchQueue.main.async { [weak self] in
-                    self?.state = .stopped
-                    LoggerProxy.debug("Warning: scanBlocksOperation operation cancelled")
-                }
-                return
-            }
-        }
-        
-        scanBlocksOperation.errorHandler = { [weak self] error in
-            DispatchQueue.main.async { [weak self] in
-                guard let self = self else { return }
-                self.processingError = error
-                self.fail(error)
-            }
-        }
-        
-        let enhanceOperation = CompactBlockEnhancementOperation(
-            rustWelding: rustBackend,
-            dataDb: config.dataDb,
-            downloader: downloader,
-            repository: transactionRepository,
-            range: range.blockRange(),
-            networkType: self.config.network.networkType
-        )
-        
-        enhanceOperation.startedHandler = {
-            LoggerProxy.debug("Started Enhancing range: \(range)")
-            DispatchQueue.main.async { [weak self] in
-                self?.state = .enhancing
-            }
-        }
-        
-        enhanceOperation.txFoundHandler = { [weak self] txs, range in
-            self?.notifyTransactions(txs, in: range)
-        }
-        
-        enhanceOperation.completionHandler  = { _, cancelled in
-            guard !cancelled else {
-                LoggerProxy.debug("Warning: enhance operation on range \(range) cancelled")
-                return
-            }
-        }
-        
-        enhanceOperation.errorHandler = { [weak self] error in
-            DispatchQueue.main.async { [weak self] in
-                guard let self = self else { return }
-                
-                self.processingError = error
-                self.fail(error)
-            }
-        }
-        
-        let scanEnhanceAdapterOperation = BlockOperation { [weak enhanceOperation, weak scanBlocksOperation] in
-            enhanceOperation?.error = scanBlocksOperation?.error
-        }
-        
-        let fetchOperation = FetchUnspentTxOutputsOperation(
-            accountRepository: accountRepository,
-            downloader: self.downloader,
-            rustbackend: rustBackend,
-            dataDb: config.dataDb,
-            startHeight: config.walletBirthday,
-            networkType: self.config.network.networkType
-        )
-        
-        fetchOperation.startedHandler = { [weak self] in
-            DispatchQueue.main.async { [weak self] in
-                self?.state = .fetching
-            }
-        }
-        
-        fetchOperation.completionHandler = {  [weak self] _, cancelled in
-            guard !cancelled else {
-                LoggerProxy.debug("Warning: fetch operation on range \(range) cancelled")
-                return
-            }
-            DispatchQueue.main.async { [weak self] in
-                self?.processBatchFinished(range: range)
-            }
-        }
-
-        fetchOperation.errorHandler = { [weak self] error in
-            DispatchQueue.main.async { [weak self] in
-                guard let self = self else { return }
-                
-                self.processingError = error
-                self.fail(error)
-            }
-        }
-
-        fetchOperation.fetchedUTXOsHandler = { result in
-            NotificationCenter.default.post(
-                name: .blockProcessorStoredUTXOs,
-                object: self,
-                userInfo: [CompactBlockProcessorNotificationKey.refreshedUTXOs: result]
-            )
-        }
-        
-        let enhanceFetchAdapterOperation = BlockOperation { [weak fetchOperation, weak enhanceOperation] in
-            fetchOperation?.error = enhanceOperation?.error
-        }
-        
-        downloadValidateAdapterOperation.addDependency(downloadBlockOperation)
-        validateChainOperation.addDependency(downloadValidateAdapterOperation)
-        validateScanningAdapterOperation.addDependency(validateChainOperation)
-        scanBlocksOperation.addDependency(validateScanningAdapterOperation)
-        scanEnhanceAdapterOperation.addDependency(scanBlocksOperation)
-        enhanceOperation.addDependency(scanEnhanceAdapterOperation)
-        enhanceFetchAdapterOperation.addDependency(enhanceOperation)
-        fetchOperation.addDependency(enhanceFetchAdapterOperation)
-        
-        operationQueue.addOperations(
-            [
-                downloadBlockOperation,
-                downloadValidateAdapterOperation,
-                validateChainOperation,
-                validateScanningAdapterOperation,
-                scanBlocksOperation,
-                scanEnhanceAdapterOperation,
-                enhanceOperation,
-                enhanceFetchAdapterOperation,
-                fetchOperation
-            ],
-            waitUntilFinished: false
-        )
     }
     
     func calculateProgress(start: BlockHeight, current: BlockHeight, latest: BlockHeight) -> Float {
@@ -892,35 +694,35 @@ public class CompactBlockProcessor {
     }
 
     func severeFailure(_ error: Error) {
-        operationQueue.cancelAllOperations()
+        cancelableTask?.cancel()
         LoggerProxy.error("show stoppper failure: \(error)")
         self.backoffTimer?.invalidate()
         self.retryAttempts = config.retries
         self.processingError = error
-        self.state = .error(error)
+        setState(.error(error))
         self.notifyError(error)
     }
 
     func fail(_ error: Error) {
         // todo specify: failure
         LoggerProxy.error("\(error)")
-        operationQueue.cancelAllOperations()
+        cancelableTask?.cancel()
         self.retryAttempts += 1
         self.processingError = error
-        switch self.state {
+        switch self.state.getState() {
         case .error:
             notifyError(error)
         default:
             break
         }
-        self.state = .error(error)
+        setState(.error(error))
         guard self.maxAttemptsReached else { return }
         // don't set a new timer if there are no more attempts.
         self.setTimer()
     }
 
     func retryProcessing(range: CompactBlockRange) {
-        operationQueue.cancelAllOperations()
+        cancelableTask?.cancel()
         // update retries
         self.retryAttempts += 1
         self.processingError = nil
@@ -935,7 +737,7 @@ public class CompactBlockProcessor {
 
             // process next batch
             // processNewBlocks(range: Self.nextBatchBlockRange(latestHeight: latestBlockHeight, latestDownloadedHeight: try downloader.lastDownloadedBlockHeight(), walletBirthday: config.walletBirthday))
-            nextBatchTask()
+            nextBatch()
         } catch {
             self.fail(error)
         }
@@ -972,48 +774,9 @@ public class CompactBlockProcessor {
         }
     }
 
-    @available(*, deprecated, message: "This static method will be removed soon, use `nextBatchTask()` instead.")
     private func nextBatch() {
-        self.state = .downloading
-        NextStateHelper.nextState(
-            service: self.service,
-            downloader: self.downloader,
-            config: self.config,
-            rustBackend: self.rustBackend,
-            queue: nil
-        ) { result in
-            DispatchQueue.main.async { [weak self] in
-                guard let self = self else { return }
-                switch result {
-                case .success(let nextState):
-                    switch nextState {
-                    case .finishProcessing(let height):
-                        self.latestBlockHeight = height
-                        self.processingFinished(height: height)
-                    case .processNewBlocks(let range):
-                        self.latestBlockHeight = range.upperBound
-                        self.lowerBoundHeight = range.lowerBound
-                        self.processNewBlocks(range: range)
-                    case let .wait(latestHeight, latestDownloadHeight):
-                        // Lightwalletd might be syncing
-                        self.lowerBoundHeight = latestDownloadHeight
-                        self.latestBlockHeight = latestHeight
-                        LoggerProxy.info(
-                            "Lightwalletd might be syncing: latest downloaded block height is: \(latestDownloadHeight)" +
-                            "while latest blockheight is reported at: \(latestHeight)"
-                        )
-                        self.processingFinished(height: latestDownloadHeight)
-                    }
-                case .failure(let error):
-                    self.severeFailure(error)
-                }
-            }
-        }
-    }
-    
-    private func nextBatchTask() {
-        self.state = .downloading
-        Task { [self] in
+        setState(.downloading)
+        Task { @MainActor [self] in
             do {
                 let nextState = try await NextStateHelper.nextStateAsync(
                     service: self.service,
@@ -1045,9 +808,9 @@ public class CompactBlockProcessor {
         }
     }
 
-    private func validationFailed(at height: BlockHeight) {
+    internal func validationFailed(at height: BlockHeight) {
         // cancel all Tasks
-        operationQueue.cancelAllOperations()
+        cancelableTask?.cancel()
 
         // register latest failure
         self.lastChainValidationFailure = height
@@ -1078,13 +841,13 @@ public class CompactBlockProcessor {
             )
             
             // process next batch
-            self.nextBatchTask()
+            self.nextBatch()
         } catch {
             self.fail(error)
         }
     }
 
-    private func processBatchFinished(range: CompactBlockRange) {
+    internal func processBatchFinished(range: CompactBlockRange) {
         guard processingError == nil else {
             retryProcessing(range: range)
             return
@@ -1098,7 +861,7 @@ public class CompactBlockProcessor {
             return
         }
         
-        nextBatchTask()
+        nextBatch()
     }
     
     private func processingFinished(height: BlockHeight) {
@@ -1110,7 +873,7 @@ public class CompactBlockProcessor {
                 CompactBlockProcessorNotificationKey.foundBlocks: self.foundBlocks
             ]
         )
-        self.state = .synced
+        setState(.synced)
         setTimer()
     }
     
@@ -1297,33 +1060,27 @@ extension UnifiedAddressShim: UnifiedAddress {
 }
 
 extension CompactBlockProcessor {
-    func refreshUTXOs(tAddress: String, startHeight: BlockHeight, result: @escaping (Result<RefreshedUTXOs, Error>) -> Void) {
+    func refreshUTXOs(tAddress: String, startHeight: BlockHeight) async throws -> RefreshedUTXOs {
         let dataDb = self.config.dataDb
-        self.downloader.fetchUnspentTransactionOutputs(tAddress: tAddress, startHeight: startHeight) { [weak self] fetchResult in
-            switch fetchResult {
-            case .success(let utxos):
-                DispatchQueue.main.async {
-                    self?.operationQueue.addOperation { [self] in
-                        guard let self = self else { return }
-                        do {
-                            guard try self.rustBackend.clearUtxos(
-                                dbData: dataDb,
-                                address: tAddress,
-                                sinceHeight: startHeight - 1,
-                                networkType: self.config.network.networkType
-                            ) >= 0 else {
-                                result(.failure(CompactBlockProcessorError.generalError(message: "attempted to clear utxos but -1 was returned")))
-                                return
-                            }
-                        } catch {
-                            result(.failure(self.mapError(error)))
-                        }
-                        result(.success(self.storeUTXOs(utxos, in: dataDb)))
-                    }
-                }
-            case .failure(let error):
-                result(.failure(self?.mapError(error) ?? error))
+        
+        let stream: AsyncThrowingStream<UnspentTransactionOutputEntity, Error> = downloader.fetchUnspentTransactionOutputs(tAddress: tAddress, startHeight: startHeight)
+        var utxos: [UnspentTransactionOutputEntity] = []
+        
+        do {
+            for try await utxo in stream {
+                utxos.append(utxo)
             }
+            guard try rustBackend.clearUtxos(
+                dbData: dataDb,
+                address: tAddress,
+                sinceHeight: startHeight - 1,
+                networkType: self.config.network.networkType
+            ) >= 0 else {
+                throw CompactBlockProcessorError.generalError(message: "attempted to clear utxos but -1 was returned")
+            }
+            return storeUTXOs(utxos, in: dataDb)
+        } catch {
+            throw mapError(error)
         }
     }
     
@@ -1386,6 +1143,7 @@ extension CompactBlockProcessorError: LocalizedError {
         case let .wrongConsensusBranchId(expectedLocally, found):
             // swiftlint:disable:next line_length
             return "The remote server you are connecting to is publishing a different branch ID \(found) than the one your App is expecting to be (\(expectedLocally)). This could be caused by your App being out of date or the server you are connecting you being either on a different network or out of date after a network upgrade."
+        case .unknown: return "Unknown error occured."
         }
     }
 
@@ -1405,12 +1163,6 @@ extension CompactBlockProcessorError: LocalizedError {
     }
 }
 
-extension CompactBlockProcessor: CompactBlockProgressDelegate {
-    func progressUpdated(_ progress: CompactBlockProgress) {
-        notifyProgress(progress)
-    }
-}
-
 extension CompactBlockProcessor: EnhancementStreamDelegate {
     func transactionEnhancementProgressUpdated(_ progress: EnhancementProgress) {
         NotificationCenter.default.post(
@@ -1423,38 +1175,12 @@ extension CompactBlockProcessor: EnhancementStreamDelegate {
 
 extension CompactBlockProcessor {
     enum NextStateHelper {
-        // swiftlint:disable:next function_parameter_count
-        static func nextState(
-            service: LightWalletService,
-            downloader: CompactBlockDownloading,
-            config: Configuration,
-            rustBackend: ZcashRustBackendWelding.Type,
-            queue: DispatchQueue?,
-            result: @escaping (Result<FigureNextBatchOperation.NextState, Error>) -> Void
-        ) {
-            let dispatchQueue = queue ?? DispatchQueue.global(qos: .userInitiated)
-
-            dispatchQueue.async {
-                do {
-                    let nextResult = try self.nextState(
-                        service: service,
-                        downloader: downloader,
-                        config: config,
-                        rustBackend: rustBackend
-                    )
-                    result(.success(nextResult))
-                } catch {
-                    result(.failure(error))
-                }
-            }
-        }
-        
         static func nextStateAsync(
             service: LightWalletService,
             downloader: CompactBlockDownloading,
             config: Configuration,
             rustBackend: ZcashRustBackendWelding.Type
-        ) async throws -> FigureNextBatchOperation.NextState {
+        ) async throws -> NextState {
             let task = Task(priority: .userInitiated) {
                 // TODO: refactor to async call, issue 463, PR 493
                 // https://github.com/zcash/ZcashLightClientKit/issues/463
@@ -1473,7 +1199,7 @@ extension CompactBlockProcessor {
             downloader: CompactBlockDownloading,
             config: Configuration,
             rustBackend: ZcashRustBackendWelding.Type
-        ) throws -> FigureNextBatchOperation.NextState {
+        ) throws -> NextState {
             let info = try service.getInfo()
             
             try CompactBlockProcessor.validateServerInfo(
