+++ Sources/ZcashLightClientKit/Block/Processor/CompactBlockValidationInformation.swift
@@ -8,78 +8,58 @@
 
 import Foundation
 
-enum CompactBlockValidationError: Error {
-    case validationFailed(height: BlockHeight)
-    case failedWithError(_ error: Error?)
-}
-class CompactBlockValidationOperation: ZcashOperation {
-    override var isConcurrent: Bool { false }
-    
-    override var isAsynchronous: Bool { false }
-    
-    var rustBackend: ZcashRustBackendWelding.Type
-    
-    private var cacheDb: URL
-    private var dataDb: URL
-    private var network: NetworkType
-    private var cancelableTask: Task<Void, Error>?
-    private var done = false
-
-    init(
-        rustWelding: ZcashRustBackendWelding.Type,
-        cacheDb: URL,
-        dataDb: URL,
-        networkType: NetworkType
-    ) {
-        rustBackend = rustWelding
-        self.cacheDb = cacheDb
-        self.dataDb = dataDb
-        self.network = networkType
-        super.init()
+extension CompactBlockProcessor {
+    enum CompactBlockValidationError: Error {
+        case validationFailed(height: BlockHeight)
+        case failedWithError(_ error: Error?)
     }
-    
-    override func main() {
-        guard !shouldCancel() else {
-            cancel()
-            return
-        }
 
-        self.startedHandler?()
+    func compactBlockValidation() async throws {
+        try Task.checkCancellation()
+        
+        setState(.validating)
 
-        cancelableTask = Task {
-            let result = self.rustBackend.validateCombinedChain(dbCache: cacheDb, dbData: dataDb, networkType: self.network)
-            
+        let result = rustBackend.validateCombinedChain(dbCache: config.cacheDb, dbData: config.dataDb, networkType: config.network.networkType)
+        
+        do {
             switch result {
             case 0:
                 let error = CompactBlockValidationError.failedWithError(rustBackend.lastError())
-                self.error = error
-                LoggerProxy.debug("block scanning failed with error: \(String(describing: self.error))")
-                self.fail(error: error)
+                LoggerProxy.debug("block scanning failed with error: \(String(describing: error))")
+                throw error
                 
             case ZcashRustBackendWeldingConstants.validChain:
-                self.done = true
+                if Task.isCancelled {
+                    setState(.stopped)
+                    LoggerProxy.debug("Warning: compactBlockValidation cancelled")
+                }
+                LoggerProxy.debug("validateChainFinished")
                 break
                 
             default:
                 let error = CompactBlockValidationError.validationFailed(height: BlockHeight(result))
-                self.error = error
-                LoggerProxy.debug("block scanning failed with error: \(String(describing: self.error))")
-                self.fail(error: error)
+                LoggerProxy.debug("block scanning failed with error: \(String(describing: error))")
+                throw error
+            }
+        } catch {
+            guard let validationError = error as? CompactBlockValidationError else {
+                LoggerProxy.error("Warning: compactBlockValidation returning generic error: \(error)")
+                return
+            }
+            
+            switch validationError {
+            case .validationFailed(let height):
+                LoggerProxy.debug("chain validation at height: \(height)")
+                validationFailed(at: height)
+            case .failedWithError(let err):
+                guard let validationFailure = err else {
+                    LoggerProxy.error("validation failed without a specific error")
+                    self.fail(CompactBlockProcessorError.generalError(message: "validation failed without a specific error"))
+                    return
+                }
+                
+                throw validationFailure
             }
         }
-        
-        while !done && !isCancelled {
-            sleep(1)
-        }
-    }
-    
-    override func fail(error: Error? = nil) {
-        self.cancelableTask?.cancel()
-        super.fail(error: error)
-    }
-    
-    override func cancel() {
-        self.cancelableTask?.cancel()
-        super.cancel()
     }
 }
